% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-country.r
\name{.get_ctry_categories}
\alias{.get_ctry_categories}
\title{Get all categories that contain specific countries}
\usage{
.get_ctry_categories(country_codes)
}
\arguments{
\item{country_codes}{Character vector of ISO3 country codes}
}
\value{
Character vector of category names
}
\description{
Get all categories that contain specific countries
}
\keyword{internal}
