% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_get_category.r
\name{im_get_category}
\alias{im_get_category}
\title{Get Countries Belonging to Specified Categories}
\usage{
im_get_category(input, verbose = TRUE)
}
\arguments{
\item{input}{Character vector of category codes or ISO3 country codes. Can include
exclusion patterns (e.g., "EUROPE-DEU" for European countries excluding Germany).}

\item{verbose}{Logical, whether to print reminder messages about specific countries
(default: TRUE)}
}
\value{
Character vector of ISO3 country codes belonging to the specified categories
}
\description{
Returns ISO3 codes for countries belonging to specified categories.
Supports complex queries including category exclusions.
}
\details{
Categories include:
\itemize{
\item Economic: CTR_LDR, CTR_FOL, SMP_WLD, SMP_RLD, SMP_FOL, PERI
\item Regional: AFRICA, AMERICAS, ASIA, EUROPE, OCEANIA
\item Subregional: NORTH_AFRICA, WEST_ASIA, etc.
\item Groups: BRICS, EU, ASEAN, etc.
\item Resources: NRS_REN, HYD_EXP, HYD_IMP
}
}
\examples{
# Get all center countries
center_countries <- im_get_category("CTR")

# Get European countries excluding Germany
europe_except_germany <- im_get_category("EUROPE-DEU")

# Get multiple categories
brics_and_eu <- im_get_category(c("BRICS", "EU"))

}
\seealso{
\code{\link{im_from_iso}} for getting specific information about countries,
\code{\link{im_ctry2iso}} for converting country names to ISO codes

Other country classification functions: 
\code{\link{im_from_iso}()}
}
\concept{country classification functions}
